chore: improve the docstrings
